refactor: provider & dependency layer (#127)
Five independent improvements to the provider/dependency layer, all
landed together:

- Findings 7: split embedder.ts into src/embedders/{types,local,remote,cache}.ts
  and keep embedder.ts as a backward-compatible facade. LocalEmbedder
  pipeline lifecycle is now an instance variable so tests can construct
  fresh embedders without leaking module-level state.
- Finding 8: remove the mutable ASSET_TYPES export. Callers now use
  getAssetTypes() (readonly) and a new deregisterAssetType() helper for
  test cleanup.
- Finding 9: rename local-search.ts -> db-search.ts to clarify this is
  the DB-layer search implementation, not a "local vs. remote" split.
  No code moves; only the file rename and import paths.
- Finding 23: add src/providers/index.ts barrel mirroring
  stash-providers/index.ts so registry-search.ts uses a single side-
  effect import for all built-in registry providers.
- Finding 27: split llm.ts into llm-client.ts (transport) and
  metadata-enhance.ts (workflow). llm.ts is now a re-export facade so
  existing imports keep working unchanged.

The previously-fixed circular import (stash-search -> filesystem
provider -> stash-search) stays fixed: filesystem.ts continues to
import searchLocal from db-search.ts, which has no edge back into
stash-search.ts.
