+++ src/embedders/remote.ts
@@ -1,0 +1,148 @@
+/**
+ * OpenAI-compatible remote embedder.
+ *
+ * Calls the configured `/embeddings` endpoint and L2-normalizes the returned
+ * vectors so the scoring pipeline's L2-to-cosine conversion is correct.
+ */
+
+import { fetchWithTimeout, isHttpUrl } from "../common";
+import type { EmbeddingConnectionConfig } from "../config";
+import type { Embedder, EmbeddingVector } from "./types";
+
+const REMOTE_BATCH_SIZE = 100;
+
+export class RemoteEmbedder implements Embedder {
+  constructor(private readonly config: EmbeddingConnectionConfig) {}
+
+  async embed(text: string): Promise<EmbeddingVector> {
+    const headers = this.buildHeaders();
+    const body: { input: string; model: string; dimensions?: number } = {
+      input: text,
+      model: this.config.model,
+    };
+    if (this.config.dimension) {
+      body.dimensions = this.config.dimension;
+    }
+
+    const response = await fetchWithTimeout(normalizeEmbeddingEndpoint(this.config.endpoint), {
+      method: "POST",
+      headers,
+      body: JSON.stringify(body),
+    });
+
+    if (!response.ok) {
+      const errBody = await response.text().catch(() => "");
+      throw new Error(`Embedding request failed (${response.status}): ${errBody}`);
+    }
+
+    const json = (await response.json()) as { data: Array<{ embedding: number[] }> };
+
+    if (!json.data?.[0]?.embedding) {
+      throw new Error(
+        `Unexpected embedding response format: missing data[0].embedding.${embeddingEndpointPathHint(this.config.endpoint)}`,
+      );
+    }
+
+    return l2Normalize(json.data[0].embedding);
+  }
+
+  async embedBatch(texts: string[]): Promise<EmbeddingVector[]> {
+    if (texts.length === 0) return [];
+    const results: EmbeddingVector[] = [];
+    const headers = this.buildHeaders();
+
+    for (let i = 0; i < texts.length; i += REMOTE_BATCH_SIZE) {
+      const batch = texts.slice(i, i + REMOTE_BATCH_SIZE);
+      const body: { input: string[]; model: string; dimensions?: number } = {
+        input: batch,
+        model: this.config.model,
+      };
+      if (this.config.dimension) {
+        body.dimensions = this.config.dimension;
+      }
+
+      const response = await fetchWithTimeout(normalizeEmbeddingEndpoint(this.config.endpoint), {
+        method: "POST",
+        headers,
+        body: JSON.stringify(body),
+      });
+
+      if (!response.ok) {
+        const respBody = await response.text().catch(() => "");
+        throw new Error(`Embedding batch request failed (${response.status}): ${respBody}`);
+      }
+
+      const json = (await response.json()) as {
+        data: Array<{ embedding: number[]; index: number }>;
+      };
+
+      if (!json.data || json.data.length !== batch.length) {
+        throw new Error(
+          `Unexpected embedding batch response: expected ${batch.length} embeddings, got ${json.data?.length ?? 0}.${embeddingEndpointPathHint(this.config.endpoint)}`,
+        );
+      }
+
+      // Sort by index to guarantee correct order (OpenAI API doesn't guarantee order)
+      const sorted = [...json.data].sort((a, b) => a.index - b.index);
+
+      for (const [idx, d] of sorted.entries()) {
+        if (!Array.isArray(d.embedding)) {
+          throw new Error(`Unexpected embedding at batch index ${idx}: missing or invalid`);
+        }
+        results.push(l2Normalize(d.embedding));
+      }
+    }
+
+    return results;
+  }
+
+  private buildHeaders(): Record<string, string> {
+    const headers: Record<string, string> = { "Content-Type": "application/json" };
+    if (this.config.apiKey) {
+      headers.Authorization = `Bearer ${this.config.apiKey}`;
+    }
+    return headers;
+  }
+}
+
+/**
+ * L2-normalize a vector to unit length.
+ * Required for remote embeddings because the scoring pipeline's L2-to-cosine
+ * conversion formula (1 - distance^2/2) is only correct for unit vectors.
+ * The local embedder already normalizes via `normalize: true`.
+ */
+function l2Normalize(vec: number[]): number[] {
+  const norm = Math.sqrt(vec.reduce((sum, v) => sum + v * v, 0));
+  if (norm === 0) return vec;
+  return vec.map((v) => v / norm);
+}
+
+export function normalizeEmbeddingEndpoint(endpoint: string): string {
+  let parsed: URL;
+  try {
+    parsed = new URL(endpoint);
+  } catch {
+    return endpoint;
+  }
+
+  const normalizedPath = parsed.pathname.replace(/\/+$/, "");
+  if (normalizedPath.endsWith("/embeddings")) {
+    return parsed.toString();
+  }
+
+  parsed.pathname = normalizedPath ? `${normalizedPath}/embeddings` : "/embeddings";
+  return parsed.toString();
+}
+
+function embeddingEndpointPathHint(endpoint: string): string {
+  const normalizedEndpoint = normalizeEmbeddingEndpoint(endpoint);
+  if (normalizedEndpoint !== endpoint) {
+    return ` Check that your endpoint includes the full embeddings path (for example "${normalizedEndpoint}", not just "${endpoint}").`;
+  }
+  return "";
+}
+
+/** Check whether an EmbeddingConnectionConfig has a valid remote endpoint. */
+export function hasRemoteEndpoint(config: EmbeddingConnectionConfig): boolean {
+  return isHttpUrl(config.endpoint);
+}
