+++ src/providers/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Centralized registry provider registration.
+ *
+ * Import this module (side-effect import) to register all built-in registry
+ * providers with the provider registry. This replaces the individual
+ * side-effect imports that were duplicated in registry-search.ts.
+ *
+ * Mirrors the pattern used by `stash-providers/index.ts`.
+ */
+import "./static-index";
+import "./skills-sh";
